Fix in add into. Removing translate to resizing mask = false from container view: is not necessary and it made load fail when the container view was the roo tview of a controller
